requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
mysqlclient==2.0.2
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
